1.0.0.26 RC8:Introduce FX11-aug2022(C++17).

+++ Effects11/EffectVariable.inl
@@ -4,7 +4,7 @@
 // Direct3D 11 Effects Variable reflection template
 // These templates define the many Effect variable types.
 //
-// Copyright (c) Microsoft Corporation. All rights reserved.
+// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 //
 // http://go.microsoft.com/fwlink/p/?LinkId=271568
@@ -619,8 +619,8 @@ inline bool AreBoundsValid(_In_ uint32_t Offset, _In_ uint32_t Count, _In_ const
 
     return ((Offset + Count >= Offset) &&
         ((Offset + Count) < ((uint32_t)-1) / pType->Stride) &&
-        ((size_t)Count * pType->Stride + (uint8_t*)pData >= (uint8_t*)pData) &&
-        (((size_t)Offset + Count - 1) * pType->Stride + singleElementSize <= TotalUnpackedSize));
+        (Count * pType->Stride + (uint8_t*)pData >= (uint8_t*)pData) &&
+        ((Offset + Count - 1) * pType->Stride + singleElementSize <= TotalUnpackedSize));
 }
 
 // Note that the branches in this code is based on template parameters and will be compiled out
